pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "brawlstar_project"
version = "0.2.0"
description = "Brawl Stars data analysis project using Polars"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "duckdb>=1.3.2",
    "isort>=6.0.1",
    "jupyterlab>=4.4.4",
    "polars>=1.31.0",
    "pre-commit>=4.2.0",
    "pydantic>=2.11.7",
    "pyiceberg>=0.9.1",
    "pytest>=8.4.1",
    "pytest-cov>=6.0.0",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
    "requests>=2.32.4",
    "ruff>=0.12.3",
    "streamlit>=1.46.1",
]

[project.scripts]
brawlstars = "brawlstar_project.processing.raw.main:main"

[tool.hatch.build.targets.wheel]
packages = ["brawlstar_project"]

[tool.hatch.build]
source-directory = "src"

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.ruff]
line-length = 88
